phonebook

+++ CPP_00/ex00/megaphone.cpp
@@ -6,7 +6,7 @@
 /*   By: diwalaku <diwalaku@codam.student.nl>         +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2025/01/16 16:42:36 by diwalaku      #+#    #+#                 */
-/*   Updated: 2025/01/16 17:25:31 by diwalaku      ########   odam.nl         */
+/*   Updated: 2025/03/05 17:42:32 by diwalaku      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -34,7 +34,7 @@ int	main(int argc, char **argv)
 		for (int i = 1; i < argc; i++)
 		{
 			std::string str = argv[i];
-			for (int j = 0; j < str.length(); j++)
+			for (size_t j = 0; j < str.length(); j++)
 				std::cout << static_cast<char>(std::toupper(str[j]));
 		}
 		std::cout << std::endl;
